Refactor LandRepository module location and imports
Renamed the LandRepository module to `land` and updated its import in `views.py` to reflect the new structure. This change improves the clarity and consistency of the repository module naming convention.
